requirements.txt
aiogram==3.13.1
httpx==0.27.0
python-dotenv==1.0.1
orjson==3.10.7

psycopg[binary]==3.2.9
psycopg-pool==3.2.4
